this is to change the colors to work more with the deliverables
(you can change line 38 background color back to black, i thought seemed more like a big room or map with background as white)

would want to change: it seems that it costs a character a turn to go towards the edge or out of the map given when they dont move compared to when they hit a wall and it doesnt cost a turn.

would want to change maybe?